Redesigned the method of obtaining a token. Added an additional check when updating the token. Reworked sending data via swagger.

+++ site_vote_photo/vote_photo/api/view/refresh_token_view.py
@@ -1,18 +1,28 @@
 from django.core.exceptions import ObjectDoesNotExist, ValidationError
 from django.contrib.auth.hashers import check_password
+from rest_framework.authentication import TokenAuthentication
+from rest_framework.parsers import MultiPartParser
+from rest_framework.permissions import IsAuthenticated
 from rest_framework.views import APIView
 from rest_framework.authtoken.models import Token
 from rest_framework.response import Response
 from rest_framework import status
+from drf_yasg.utils import swagger_auto_schema
 
 from vote_photo.mymodels.model_user import User
+from ..serializers.refresh_token_serializers import RefreshTokenSerializers
 
 
 class RefreshTokenView(APIView):
+    permission_classes = (IsAuthenticated,)
+    authentication_classes = (TokenAuthentication,)
+    parser_classes = (MultiPartParser,)
+
+    @swagger_auto_schema(request_body=RefreshTokenSerializers)
     def put(self, request, format=None):
         try:
-            user = User.objects.get(username=request.data["username"])
-            if check_password(request.data["password"], user.password):
+            user_get_request = User.objects.get(username=request.data["username"])
+            if check_password(request.data["password"], user_get_request.password):
                 pass
             else:
                 raise ValidationError("Не верный пароль.")
@@ -35,7 +45,7 @@ class RefreshTokenView(APIView):
             )
         try:
             token = Token.objects.get(
-                user=user,
+                user=user_get_request.id,
             )
         except ObjectDoesNotExist:
             return Response(
@@ -45,14 +55,23 @@ class RefreshTokenView(APIView):
                 },
                 status=404,
             )
-        token.delete()
-        return Response(
-            {
-                "new_token": Token.objects.create(
-                    user=User.objects.get(username=request.data["username"])
-                ).key,
-                "username": request.data["username"],
-                "status": status.HTTP_200_OK,
-            },
-            status=status.HTTP_200_OK,
-        )
+        if request.user == user_get_request:
+            token.delete()
+            return Response(
+                {
+                    "new_token": Token.objects.create(
+                        user=User.objects.get(username=request.data["username"])
+                    ).key,
+                    "username": request.data["username"],
+                    "status": status.HTTP_200_OK,
+                },
+                status=status.HTTP_200_OK,
+            )
+        else:
+            return Response(
+                {
+                    "error": f"Вы не можете изменить токен '{user_get_request.username}'.",
+                    "status": status.HTTP_400_BAD_REQUEST,
+                },
+                status=status.HTTP_400_BAD_REQUEST,
+            )
